functions/ingestor/requirements.txt
# Function dependencies, for example:
# package>=version
fsspec
gcsfs
pandas
pyarrow
google.cloud.bigquery
google.cloud.storage